add password and email validation to Registration component
